Correctly delete participants from assignment.   * If the participant is associated with a review map (or responses) or a team, the user is prompted     before deleting the item. The user is given three options:       1) Delete the participant and its associated items       2) Display a list of associated items and selectively delete zero or more of the items       3) Do nothing   * If the user chooses the second option and deletes all the associated items, the participant will be removed automatically. If any items remain, the user will be given the prompt again for another action.   * If a participant is the only member of a team when he or she is deleted, the team is also removed.
git-svn-id: https://svn.csc.ncsu.edu:8443/svn/research-expertiza/trunk@762 392d7b7b-3f31-0410-9dc3-c95a9635ea79

+++ app/models/participant.rb
@@ -12,15 +12,41 @@ class Participant < ActiveRecord::Base
   def fullname
     User.find(self.user_id).fullname
   end 
-
-  def delete
-    times = ResubmissionTime.find(:all, :conditions => ['participant_id = ?',self.id])
-    times.each {|time| time.destroy }
-    begin
-      self.destroy
-    rescue
-      puts self.id
+  
+  def delete(force = nil)     
+    maps = ResponseMap.find(:all, :conditions => ['reviewee_id = ? or reviewer_id = ?',self.id,self.id])
+    
+    if force or ((maps.nil? or maps.length == 0) and 
+                 self.team.nil?)
+      force_delete(maps)
+    else
+      raise "Associations exist for this participant"        
     end
+  end
+  
+  def force_delete(maps)
+    times = ResubmissionTime.find(:first, :conditions => ['participant_id = ?',self.id])    
+    
+    if times
+      times.each { |time| time.destroy }
+    end
+    
+    if maps
+      maps.each { |map| map.delete(true) }
+    end
+    
+    if self.team
+      if self.team.teams_users.length == 1
+        self.team.delete
+      else
+        self.team.teams_users.each{ |tuser| 
+          if tuser.user_id == self.id
+            tuser.delete
+          end
+        }
+      end
+     end
+     self.destroy
   end
 
   def get_topic_string
